requirements.txt
fastapi
uvicorn
python-multipart
assemblyai
murf
google-generativeai
